requirements.txt
streamlit 
langchain-openai 
langchain-groq 
langchain-google-genai 
langchain-core
sentence-transformers
faiss-cpu
python-dotenv
requests




